Added replay plus other stuff that I don't remember

+++ core/src/main/scala/org/apache/spark/TaskContextImpl.scala
@@ -48,7 +48,7 @@ private[spark] class TaskContextImpl(val stageId: Int,
   @transient var currentInputId: Int = -1
 
   // Used to pipeline records through taps inside the same stage
-  @transient var currentBuffer: ByteBuffer[Long, Int] = null
+  @transient var currentBuffer: ByteBuffer[Int, Int] = null
 
   @transient var threadPool: ThreadPoolExecutor = null
